changed example to clear all for combobox
