+++ site/docs/components/combo-box/examples.mdx
@@ -178,11 +178,11 @@ Use the `truncate` prop when you are limited on space and want to prevent the mu
 
 </LivePreview>
 
-<LivePreview componentName="combo-box" exampleName="SelectAll" >
+<LivePreview componentName="combo-box" exampleName="ClearAll" >
 
-## Select All
+## Clear All
 
-In order to achieve the select all functionality, you can pass the first Option as Select All and pass in the logic to handleSelectionChange prop.
+In order to achieve the clear all functionality, you can pass a close icon in endAdornment prop.
 
 </LivePreview>
 </LivePreviewControls>
