+++ site/src/examples/combo-box/index.ts
@@ -14,4 +14,4 @@ export * from "./CustomFiltering";
 export * from "./ServerSideData";
 export * from "./ObjectValues";
 export * from "./Truncation";
-export * from "./SelectAll";
+export * from "./ClearAll";
